chore: refactor redis & handler adapters
