+++ collab-service/src/adapter/question_handler.ts
