+++ collab-service/src/adapter/room_handler.ts
